+++ src/lib/upload-asset.ts
@@ -1,5 +1,9 @@
-import AWS from 'aws-sdk'
-import { PutObjectRequest } from 'aws-sdk/clients/s3'
+import {
+  CompleteMultipartUploadCommandOutput,
+  PutObjectCommandInput,
+  S3Client,
+} from '@aws-sdk/client-s3'
+import { Upload } from '@aws-sdk/lib-storage'
 import { PreFillMeta } from '../types'
 
 export default async function uploadAsset<T>(
@@ -7,25 +11,39 @@ export default async function uploadAsset<T>(
   data: T,
   assetContentType?: string,
 ): Promise<{ Location: string }> {
-  const s3 = new AWS.S3({
+  const s3Client = new S3Client({
     region: assetCredentials.s3.region,
-    accessKeyId: assetCredentials.credentials.AccessKeyId,
-    secretAccessKey: assetCredentials.credentials.SecretAccessKey,
-    sessionToken: assetCredentials.credentials.SessionToken,
+    credentials: {
+      accessKeyId: assetCredentials.credentials.AccessKeyId,
+      secretAccessKey: assetCredentials.credentials.SecretAccessKey,
+      sessionToken: assetCredentials.credentials.SessionToken,
+    },
   })
 
-  const objectMeta: PutObjectRequest = {
+  const objectMeta: PutObjectCommandInput = {
     Bucket: assetCredentials.s3.bucket,
     Key: assetCredentials.s3.key,
-    Body: data as PutObjectRequest['Body'],
+    Body: data as PutObjectCommandInput['Body'],
     Expires: new Date(new Date().setFullYear(new Date().getFullYear() + 1)), // Max 1 year
     CacheControl: 'max-age=31536000', // Max 1 year(365 days),
     ContentType: assetContentType,
     ACL: 'public-read',
   }
-  const uploadOptions = {
+  const managedUpload = new Upload({
+    client: s3Client,
+    params: objectMeta,
     partSize: 5 * 1024 * 1024,
     queueSize: 1,
+    //Related github issue: https://github.com/aws/aws-sdk-js-v3/issues/2311
+    //This is a variable that is set to false by default, setting it to true
+    //means that it will force the upload to fail when one part fails on
+    //an upload. The S3 client has built in retry logic to retry uploads by default
+    leavePartsOnError: true,
+  })
+
+  const completeMultipartUploadCommandOutput: CompleteMultipartUploadCommandOutput =
+    await managedUpload.done()
+  return completeMultipartUploadCommandOutput as {
+    Location: string
   }
-  return await s3.upload(objectMeta, uploadOptions).promise()
 }